Eventstore should return valid JSON
In this commit I have added an extra guard and exception that when the
eventstore returns something other than a valid json object when you
query for the state or results that it will fail.

Previously this situation was silently ignored but for the proper handling
of eventstore failures we would want to catch these situations.